refactor(steps-2-3): extract ChimeResolver and dedup classes from monolith
Step 2: ChimeResolver class in voxd/chimes.py with _CHIME_MAP and
resolve() method. 9 new tests in test_voxd_chimes.py.

Step 3: ChimeDedup, OnceDedup, DedupHit moved to voxd/dedup.py with
associated constants. 18 tests moved to test_voxd_dedup.py with
updated monkeypatch targets.

module_size: 2330→1922 (-408), classes_per_module: 7→3 (-4),
init_violations: 3→1 (-2). Test count: 1444→1453 (+9 chime tests).
