+++ .oo-audit.jsonl
@@ -9,3 +9,8 @@
 {"ts":"2026-05-13T23:57:03Z","commit":"fe09312","files_scored":35,"files_improved":2,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/service.py":{"module_size":[990.0,978.0]},"src/punt_vox/voxd.py":{"module_size":[2342.0,2330.0]}}}
 {"ts":"2026-05-14T00:12:40Z","commit":"c0db7d6","files_scored":37,"files_improved":3,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.0],"max_complexity":[0.0,0.0],"avg_complexity":[0.0,0.0],"module_size":[0.0,102.0],"classes_per_module":[0.0,0.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.0,0.103],"encapsulation_ratio":[0.0,0.24],"avg_params":[0.0,2.11],"max_complexity":[0.0,20.0],"avg_complexity":[0.0,3.82],"module_size":[0.0,2106.0],"classes_per_module":[0.0,6.0],"class_to_func_ratio":[0.0,0.103],"init_violations":[0.0,3.0],"public_attr_violations":[0.0,19.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/config.py":{"method_ratio":[0.0,0.778],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.39],"max_complexity":[0.0,10.0],"avg_complexity":[0.0,2.11],"module_size":[0.0,275.0],"classes_per_module":[0.0,2.0],"class_to_func_ratio":[0.0,0.333],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
 {"ts":"2026-05-14T00:15:40Z","commit":"c0db7d6","files_scored":37,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"avg_params":[1.9,2.11],"avg_complexity":[3.71,3.82],"module_size":[2330.0,2106.0],"classes_per_module":[7.0,6.0]}}}
+{"ts":"2026-05-14T00:25:33Z","commit":"6223355","files_scored":39,"files_improved":3,"files_regressed":1,"verdict":"fail","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[102.0,104.0]},"src/punt_vox/voxd/chimes.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.5],"max_complexity":[0.0,4.0],"avg_complexity":[0.0,2.5],"module_size":[0.0,42.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/dedup.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.0],"max_complexity":[0.0,8.0],"avg_complexity":[0.0,3.2],"module_size":[0.0,165.0],"classes_per_module":[0.0,3.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,2.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T00:26:08Z","commit":"6223355","files_scored":39,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"avg_params":[1.9,2.23],"avg_complexity":[3.71,3.88],"module_size":[2330.0,1922.0],"classes_per_module":[7.0,3.0],"init_violations":[3.0,1.0]}}}
+{"ts":"2026-05-14T00:32:20Z","commit":"6223355","files_scored":39,"files_improved":4,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[104.0,106.0]},"src/punt_vox/voxd/_monolith.py":{"avg_params":[1.9,2.23],"avg_complexity":[3.71,3.88],"module_size":[2330.0,1922.0],"classes_per_module":[7.0,3.0],"init_violations":[3.0,1.0]},"src/punt_vox/voxd/chimes.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.5],"max_complexity":[0.0,4.0],"avg_complexity":[0.0,2.5],"module_size":[0.0,42.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/dedup.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.0],"max_complexity":[0.0,8.0],"avg_complexity":[0.0,3.2],"module_size":[0.0,165.0],"classes_per_module":[0.0,3.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,2.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T00:33:51Z","commit":"6223355","files_scored":39,"files_improved":3,"files_regressed":1,"verdict":"fail","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[102.0,106.0]},"src/punt_vox/voxd/chimes.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.5],"max_complexity":[0.0,4.0],"avg_complexity":[0.0,2.5],"module_size":[0.0,42.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/dedup.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.0],"max_complexity":[0.0,8.0],"avg_complexity":[0.0,3.2],"module_size":[0.0,165.0],"classes_per_module":[0.0,3.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,2.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T00:35:08Z","commit":"6223355","files_scored":39,"files_improved":4,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[102.0,106.0]},"src/punt_vox/voxd/_monolith.py":{"avg_params":[1.9,2.23],"avg_complexity":[3.71,3.88],"module_size":[2330.0,1922.0],"classes_per_module":[7.0,3.0],"init_violations":[3.0,1.0]},"src/punt_vox/voxd/chimes.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.5],"max_complexity":[0.0,4.0],"avg_complexity":[0.0,2.5],"module_size":[0.0,42.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/dedup.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.0],"max_complexity":[0.0,8.0],"avg_complexity":[0.0,3.2],"module_size":[0.0,165.0],"classes_per_module":[0.0,3.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,2.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
