+++ .oo-baseline.json
@@ -434,7 +434,7 @@
     "avg_params": 0.0,
     "max_complexity": 0.0,
     "avg_complexity": 0.0,
-    "module_size": 102.0,
+    "module_size": 106.0,
     "classes_per_module": 0.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
@@ -442,16 +442,29 @@
     "future_annotations": 1.0
   },
   "src/punt_vox/voxd/_monolith.py": {
-    "method_ratio": 0.103,
-    "encapsulation_ratio": 0.24,
-    "avg_params": 1.9,
+    "method_ratio": 0.019,
+    "encapsulation_ratio": 0.0,
+    "avg_params": 2.23,
     "max_complexity": 20.0,
-    "avg_complexity": 3.71,
-    "module_size": 2330.0,
-    "classes_per_module": 7.0,
-    "class_to_func_ratio": 0.103,
-    "init_violations": 3.0,
+    "avg_complexity": 3.88,
+    "module_size": 1922.0,
+    "classes_per_module": 3.0,
+    "class_to_func_ratio": 0.056,
+    "init_violations": 1.0,
     "public_attr_violations": 19.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/chimes.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 0.5,
+    "max_complexity": 4.0,
+    "avg_complexity": 2.5,
+    "module_size": 42.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
     "future_annotations": 1.0
   },
   "src/punt_vox/voxd/config.py": {
@@ -464,6 +477,19 @@
     "classes_per_module": 2.0,
     "class_to_func_ratio": 0.333,
     "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/dedup.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 1.0,
+    "max_complexity": 8.0,
+    "avg_complexity": 3.2,
+    "module_size": 165.0,
+    "classes_per_module": 3.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 2.0,
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
   },
