+++ src/punt_vox/voxd/_monolith.py
@@ -12,9 +12,7 @@ from __future__ import annotations
 import asyncio
 import base64
 import contextlib
-import hashlib
 import hmac
-import importlib.resources
 import json
 import logging
 import math
@@ -50,6 +48,7 @@ from punt_vox.types import (
     DirectPlayProvider,
     TTSProvider,
 )
+from punt_vox.voxd.chimes import ChimeResolver
 from punt_vox.voxd.config import (  # pyright: ignore[reportPrivateUsage]
     DaemonConfig,
     _config_dir,
@@ -57,6 +56,7 @@ from punt_vox.voxd.config import (  # pyright: ignore[reportPrivateUsage]
     _log_dir,
     _run_dir,
 )
+from punt_vox.voxd.dedup import ChimeDedup, OnceDedup
 
 if TYPE_CHECKING:
     from starlette.requests import Request
@@ -65,9 +65,6 @@ logger = logging.getLogger(__name__)
 
 DEFAULT_PORT = 8421
 DEFAULT_HOST = "127.0.0.1"
-
-# Audio deduplication window: skip identical audio within this many seconds.
-_DEDUP_WINDOW_SECONDS = 5.0
 
 # Lock to serialize os.environ mutation during synthesis with per-request API keys.
 _env_lock = asyncio.Lock()
@@ -451,192 +448,6 @@ async def _playback_consumer(ctx: DaemonContext) -> None:
 
 
 # ---------------------------------------------------------------------------
-# Audio deduplication
-# ---------------------------------------------------------------------------
-
-
-class ChimeDedup:
-    """Always-on in-memory dedup for chime signals.
-
-    Chimes are event markers (tests-pass, lint-fail, git-push-ok, etc)
-    and a user does not want to hear the same event chime twice in rapid
-    succession. Unlike speech, chime deduplication is always on and
-    keyed only on the signal name. The window matches the legacy
-    `AudioDedup` default so the user-visible chime behavior is
-    unchanged from versions prior to vox-0e9.
-    """
-
-    def __init__(self, window: float = _DEDUP_WINDOW_SECONDS) -> None:
-        self._window = window
-        self._seen: dict[str, float] = {}
-
-    def should_play(self, signal: str) -> bool:
-        """Return True if this chime should play (not a recent duplicate)."""
-        key = hashlib.md5(f"chime:{signal}".encode(), usedforsecurity=False).hexdigest()
-        now = time.monotonic()
-        last = self._seen.get(key)
-        if last is not None and (now - last) < self._window:
-            return False
-        self._seen[key] = now
-        cutoff = now - self._window * 2
-        self._seen = {k: v for k, v in self._seen.items() if v > cutoff}
-        return True
-
-
-@dataclass(frozen=True)
-class DedupHit:
-    """Returned when an opt-in speech dedup catches a duplicate request.
-
-    The caller made a synthesize/direct_play request with ``once=<ttl>``
-    and an identical text was already played within the TTL window. The
-    user has already heard the message — the caller should NOT treat
-    this as an error or retry. The fields support observability (logging
-    "wall skipped, already played 53s ago") and future UI surfaces.
-    """
-
-    original_played_at: float
-    """Wall-clock ``time.time()`` of the original play, in seconds since
-    the epoch. Safe to serialize, safe to compare against other wall
-    clocks. NOT ``time.monotonic()``."""
-
-    ttl_seconds_remaining: float
-    """How many more seconds this dedup key remains valid. When the
-    TTL expires, the next identical request will play fresh."""
-
-
-_ONCE_DEDUP_MAX_TTL_SECONDS: float = 3600.0
-"""Hard cap on per-call ``once`` TTL to bound memory usage. Any caller
-passing a larger value is clamped to this cap with a log warning. The
-biff wall use case targets 600 seconds; 3600 is 6x that, more than
-enough headroom without letting a stray ``once=99999999`` wedge
-``OnceDedup._seen`` with long-lived entries."""
-
-_ONCE_DEDUP_MAX_ENTRIES: int = 1024
-"""Hard cap on the number of tracked keys. Entries are evicted in
-insertion order (oldest first) when the cap is reached. Defensive
-against pathological workloads that insert thousands of unique texts
-faster than the time-based pruner can drop them."""
-
-
-class OnceDedup:
-    """Opt-in in-memory dedup for speech with per-call TTL.
-
-    Callers pass ``once=<ttl_seconds>`` on a synthesize or direct_play
-    WebSocket message (or the ``vox unmute --once <seconds>`` CLI flag)
-    to suppress duplicate plays of identical text within their chosen
-    window. Identical text spoken with different voices or providers
-    collapses — the dedup key is ``md5(text)`` only.
-
-    The motivating use case is ``biff wall``: N Claude Code sessions
-    in the same repo independently shell out to ``vox unmute`` on the
-    same broadcast text, and the user should hear the announcement
-    exactly once. See bead vox-0e9.
-
-    Unlike the legacy always-on ``AudioDedup``, this class is only
-    invoked when the caller explicitly opts in. Requests without an
-    ``once`` parameter play every time, even if identical to a recent
-    one. This preserves the property that ``vox unmute "hello"`` twice
-    in quick succession on the CLI produces two audible plays.
-
-    Per-caller TTL semantics: each caller's ``ttl_seconds`` applies to
-    THEIR query, not to the stored entry. The dedup question each
-    caller asks is "was this text played in the last ttl_seconds?" — a
-    caller passing ``once=60`` will see dedup only if the original play
-    was within 60 seconds, regardless of whether an earlier caller
-    passed ``once=600``. This matches the intuitive "dedupe within N
-    seconds" semantic.
-
-    Concurrency: ``check_and_record`` is atomic under voxd's
-    single-threaded asyncio event loop. A failed synthesis path MUST
-    call ``rollback`` to remove the zombie entry; otherwise the next
-    identical call would be incorrectly deduped against a playback
-    that never happened.
-    """
-
-    def __init__(self) -> None:
-        # key -> (inserted_monotonic, inserted_wall_clock)
-        self._seen: dict[str, tuple[float, float]] = {}
-
-    def check_and_record(self, text: str, ttl_seconds: float) -> DedupHit | None:
-        """Check for a recent duplicate; record this call if none found.
-
-        Args:
-            text: The speech text. Used as the dedup key via ``md5``.
-            ttl_seconds: Dedup window in seconds. Must be positive.
-                Values above ``_ONCE_DEDUP_MAX_TTL_SECONDS`` are
-                clamped to the cap with a log warning.
-
-        Returns:
-            ``None`` if no duplicate exists within the window — the
-            caller should proceed with synthesis + playback and call
-            ``rollback(text)`` on failure so the zombie entry is
-            removed.
-            ``DedupHit(...)`` if a duplicate was found — the caller
-            should skip the play and return the hit to its client so
-            the client can render an observable "deduped" response.
-
-        Raises:
-            ValueError: if ``ttl_seconds`` is zero or negative.
-        """
-        if ttl_seconds <= 0:
-            msg = f"ttl_seconds must be positive, got {ttl_seconds}"
-            raise ValueError(msg)
-        if ttl_seconds > _ONCE_DEDUP_MAX_TTL_SECONDS:
-            logger.warning(
-                "OnceDedup: ttl_seconds=%.1f exceeds cap %.1f, clamping",
-                ttl_seconds,
-                _ONCE_DEDUP_MAX_TTL_SECONDS,
-            )
-            ttl_seconds = _ONCE_DEDUP_MAX_TTL_SECONDS
-
-        key = hashlib.md5(text.encode(), usedforsecurity=False).hexdigest()
-        now_mono = time.monotonic()
-        now_wall = time.time()
-
-        existing = self._seen.get(key)
-        if existing is not None:
-            inserted_mono, inserted_wall = existing
-            age = now_mono - inserted_mono
-            # Per-caller semantics: the dedup fires only if the age is
-            # within THIS caller's ttl_seconds, not the stored entry's.
-            if age < ttl_seconds:
-                return DedupHit(
-                    original_played_at=inserted_wall,
-                    ttl_seconds_remaining=ttl_seconds - age,
-                )
-
-        self._seen[key] = (now_mono, now_wall)
-
-        # Opportunistic time-based prune: drop entries older than the
-        # cap so we never accumulate entries beyond the cap horizon.
-        cutoff = now_mono - _ONCE_DEDUP_MAX_TTL_SECONDS
-        self._seen = {k: (m, w) for k, (m, w) in self._seen.items() if m > cutoff}
-
-        # Hard cap on dict size. If somehow the time prune left us with
-        # more than _ONCE_DEDUP_MAX_ENTRIES, evict oldest-first. This is
-        # defensive against pathological inserts at a rate faster than
-        # the time pruner can keep up.
-        if len(self._seen) > _ONCE_DEDUP_MAX_ENTRIES:
-            sorted_items = sorted(self._seen.items(), key=lambda kv: kv[1][0])
-            keep = sorted_items[-_ONCE_DEDUP_MAX_ENTRIES:]
-            self._seen = dict(keep)
-
-        return None
-
-    def rollback(self, text: str) -> None:
-        """Remove a recorded entry for *text*, if present.
-
-        Used when synthesis or playback fails after
-        ``check_and_record`` returned None: without a rollback, the
-        zombie entry would incorrectly dedup subsequent retries
-        against a playback that never happened. Idempotent — safe
-        to call when no entry exists.
-        """
-        key = hashlib.md5(text.encode(), usedforsecurity=False).hexdigest()
-        self._seen.pop(key, None)
-
-
-# ---------------------------------------------------------------------------
 # Daemon context
 # ---------------------------------------------------------------------------
 
@@ -675,41 +486,7 @@ class DaemonContext:
         self.music_replay: bool = False
 
 
-# ---------------------------------------------------------------------------
-# Chime asset resolution
-# ---------------------------------------------------------------------------
-
-_CHIME_MAP: dict[str, str] = {
-    "done": "chime_done.mp3",
-    "prompt": "chime_prompt.mp3",
-    "acknowledge": "chime_done.mp3",
-    "compact": "chime_done.mp3",
-    "subagent": "chime_done.mp3",
-    "farewell": "chime_done.mp3",
-    "tests-pass": "chime_tests_pass.mp3",
-    "tests-fail": "chime_tests_fail.mp3",
-    "lint-pass": "chime_lint_pass.mp3",
-    "lint-fail": "chime_lint_fail.mp3",
-    "git-push-ok": "chime_git_push_ok.mp3",
-    "merge-conflict": "chime_merge_conflict.mp3",
-}
-
-
-def _resolve_chime(signal: str) -> Path | None:
-    """Resolve a chime signal name to a bundled asset path."""
-    filename = _CHIME_MAP.get(signal)
-    if filename is None:
-        return None
-    try:
-        ref = importlib.resources.files("punt_vox.assets").joinpath(filename)
-        # as_file returns a context manager; we need the actual path.
-        # For installed packages the file is already on disk.
-        path = Path(str(ref))
-        if path.exists():
-            return path
-    except (TypeError, FileNotFoundError):
-        pass
-    return None
+_chime_resolver = ChimeResolver()
 
 
 # ---------------------------------------------------------------------------
@@ -1382,7 +1159,7 @@ async def _handle_chime(
 ) -> None:
     """Handle a 'chime' message: play a bundled chime sound."""
     signal = str(msg.get("signal", "done"))
-    path = _resolve_chime(signal)
+    path = _chime_resolver.resolve(signal)
     if path is None:
         logger.warning("Unknown chime signal: %s", signal)
         await websocket.send_json(
